+++ history.md
@@ -1,3 +1,10 @@
+[2019/06/12]
+- Release: 0.0.3
+- Update
+  + open.commons.spring.jdbc.dao.AbstractGenericDao
+    - addQueryForInClause(StringBuffer, int)
+    - addQueryForInClause(StringBuffer, String, String, int)
+
 [2019/06/05]
 - Release: 0.0.2
 - Update